refactor: PostMainPageResponse record로 변경 및 Entity to DTO 로직 Converter로 변경
